ABW-1949.- Fee breakdown implemented.

+++ app/src/main/java/com/babylon/wallet/android/domain/model/AccountWithResources.kt
@@ -25,9 +25,9 @@ data class Resources(
     val isNotEmpty: Boolean
         get() = fungibleResources.isNotEmpty() || nonFungibleResources.isNotEmpty()
 
-    fun hasXrd(minimumBalance: Long = 1L): Boolean = xrd?.let {
+    fun hasXrd(minimumBalance: BigDecimal = BigDecimal(1)): Boolean = xrd?.let {
         it.amount?.let { amount ->
-            amount >= BigDecimal(minimumBalance)
+            amount >= minimumBalance
         }
     } == true
 
@@ -36,7 +36,7 @@ data class Resources(
     }
 }
 
-fun List<AccountWithResources>.findAccountWithEnoughXRDBalance(minimumBalance: Long) = find {
+fun List<AccountWithResources>.findAccountWithEnoughXRDBalance(minimumBalance: BigDecimal) = find {
     it.resources?.hasXrd(minimumBalance) ?: false
 }
